Show the update card in the multiplayer shell (self-host, Cloudflare, cloud)
The card lived only in the local/desktop shell, so self-host and Cloudflare
served the dist-tags endpoint but never rendered a card. Extract the card (npm
and desktop-native variants) into one shared SidebarUpdateCard in
@executor-js/react and drop it into both shells, so every web deployment shows
it. Add per-host render scenarios proving the card paints on the self-host /
Docker UI and on the real Cloudflare worker.
