+++ packages/app/src/web/shell.tsx
@@ -1,8 +1,6 @@
 import { Link, Outlet, useLocation } from "@tanstack/react-router";
-import { useCallback, useEffect, useRef, useState } from "react";
+import { useEffect, useRef, useState } from "react";
 import { useAtomRefresh, useAtomValue } from "@effect/atom-react";
-import * as Effect from "effect/Effect";
-import * as Exit from "effect/Exit";
 import * as AsyncResult from "effect/unstable/reactivity/AsyncResult";
 import type { Integration } from "@executor-js/sdk/shared";
 import {
@@ -11,13 +9,11 @@ import {
   toolsAllAtom,
 } from "@executor-js/react/api/atoms";
 import { Button } from "@executor-js/react/components/button";
-import { toast } from "@executor-js/react/components/sonner";
-import { copyToClipboard } from "@executor-js/react/lib/clipboard";
 import { integrationPresetIconUrl } from "@executor-js/react/components/integration-favicon";
 import { IntegrationIconWithAccount } from "@executor-js/react/components/integration-icon-with-account";
 import { CommandPalette } from "@executor-js/react/components/command-palette";
 import { useClientPlugins, useIntegrationPlugins } from "@executor-js/sdk/client";
-import { type DesktopUpdate, useDesktopUpdate } from "@executor-js/react/hooks/desktop-update";
+import { SidebarUpdateCard } from "@executor-js/react/components/update-card";
 import { ServerConnectionMenu } from "./server-connection-menu";
 
 // ── Env ─────────────────────────────────────────────────────────────────
@@ -32,247 +28,6 @@ const { VITE_APP_VERSION, VITE_GITHUB_URL } = (
     readonly env: AppMetaEnv;
   }
 ).env;
-
-// ── Version helpers ─────────────────────────────────────────────────────
-
-type UpdateChannel = "latest" | "beta";
-
-const EXECUTOR_DIST_TAGS_PATH = "/v1/app/npm/dist-tags";
-
-type ParsedVersion = {
-  readonly major: number;
-  readonly minor: number;
-  readonly patch: number;
-  readonly prerelease: ReadonlyArray<string | number> | null;
-};
-
-const semverPattern =
-  /^(?<major>\d+)\.(?<minor>\d+)\.(?<patch>\d+)(?:-(?<prerelease>[0-9A-Za-z.-]+))?(?:\+[0-9A-Za-z.-]+)?$/;
-
-const resolveUpdateChannel = (version: string): UpdateChannel =>
-  version.includes("-beta.") ? "beta" : "latest";
-
-const parseVersion = (version: string): ParsedVersion | null => {
-  const match = version.trim().match(semverPattern);
-  if (!match?.groups) return null;
-  return {
-    major: Number(match.groups.major),
-    minor: Number(match.groups.minor),
-    patch: Number(match.groups.patch),
-    prerelease: match.groups.prerelease
-      ? match.groups.prerelease.split(".").map((id) => (/^\d+$/.test(id) ? Number(id) : id))
-      : null,
-  };
-};
-
-const comparePrereleaseIdentifiers = (
-  left: ReadonlyArray<string | number> | null,
-  right: ReadonlyArray<string | number> | null,
-): number => {
-  if (left === null && right === null) return 0;
-  if (left === null) return 1;
-  if (right === null) return -1;
-  const max = Math.max(left.length, right.length);
-  for (let i = 0; i < max; i++) {
-    const l = left[i];
-    const r = right[i];
-    if (l === undefined) return -1;
-    if (r === undefined) return 1;
-    if (l === r) continue;
-    if (typeof l === "number" && typeof r === "number") return l < r ? -1 : 1;
-    if (typeof l === "number") return -1;
-    if (typeof r === "number") return 1;
-    return l < r ? -1 : 1;
-  }
-  return 0;
-};
-
-const compareVersions = (left: string, right: string): number | null => {
-  const lv = parseVersion(left);
-  const rv = parseVersion(right);
-  if (!lv || !rv) return null;
-  if (lv.major !== rv.major) return lv.major < rv.major ? -1 : 1;
-  if (lv.minor !== rv.minor) return lv.minor < rv.minor ? -1 : 1;
-  if (lv.patch !== rv.patch) return lv.patch < rv.patch ? -1 : 1;
-  return comparePrereleaseIdentifiers(lv.prerelease, rv.prerelease);
-};
-
-// ── useLatestVersion ────────────────────────────────────────────────────
-
-function useLatestVersion(currentVersion: string) {
-  const channel = resolveUpdateChannel(currentVersion);
-  const [latestVersion, setLatestVersion] = useState<string | null>(null);
-
-  useEffect(() => {
-    let cancelled = false;
-    void Effect.runPromiseExit(
-      Effect.tryPromise({
-        try: async () => {
-          const res = await fetch(EXECUTOR_DIST_TAGS_PATH);
-          if (!res.ok) return null;
-          return (await res.json()) as Partial<Record<UpdateChannel, string>>;
-        },
-        catch: (cause) => cause,
-      }),
-    ).then((exit) => {
-      if (!cancelled && Exit.isSuccess(exit)) {
-        setLatestVersion(exit.value?.[channel] ?? null);
-      }
-    });
-    return () => {
-      cancelled = true;
-    };
-  }, [channel]);
-
-  const updateAvailable =
-    latestVersion !== null && compareVersions(currentVersion, latestVersion) === -1;
-
-  return { latestVersion, updateAvailable, channel };
-}
-
-// ── UpdateCard ──────────────────────────────────────────────────────────
-
-function UpdateCard(props: { latestVersion: string; channel: UpdateChannel }) {
-  const command = `npm i -g executor@${props.channel}`;
-  const [copied, setCopied] = useState(false);
-
-  const handleCopy = useCallback(() => {
-    void copyToClipboard(command).then((ok) => {
-      if (!ok) {
-        toast.error("Failed to copy to clipboard");
-        return;
-      }
-      setCopied(true);
-      setTimeout(() => setCopied(false), 1500);
-    });
-  }, [command]);
-
-  return (
-    <div className="mx-2 mb-2 rounded-xl border border-primary/25 bg-primary/[0.06] p-3">
-      <div className="flex items-center gap-2">
-        <div className="flex size-5 shrink-0 items-center justify-center rounded-full bg-primary/15">
-          <svg viewBox="0 0 16 16" fill="none" className="size-3 text-primary">
-            <path
-              d="M8 3v7M5 7l3 3 3-3"
-              stroke="currentColor"
-              strokeWidth="1.4"
-              strokeLinecap="round"
-              strokeLinejoin="round"
-            />
-            <path d="M3 12h10" stroke="currentColor" strokeWidth="1.4" strokeLinecap="round" />
-          </svg>
-        </div>
-        <div className="min-w-0">
-          <p className="text-xs font-semibold text-foreground">Update available</p>
-          <p className="text-sm text-muted-foreground">v{props.latestVersion}</p>
-        </div>
-      </div>
-      <Button
-        type="button"
-        variant="outline"
-        onClick={handleCopy}
-        className="mt-2.5 flex w-full items-center justify-between gap-2 rounded-lg border-border/60 bg-background/50 px-2.5 py-1.5 text-left hover:bg-background/80"
-      >
-        <code className="truncate font-mono text-xs text-sidebar-foreground">{command}</code>
-        <span className="shrink-0 text-muted-foreground transition-colors group-hover:text-foreground">
-          {copied ? (
-            <svg viewBox="0 0 16 16" fill="none" className="size-3 text-primary">
-              <path
-                d="M3 8.5l3.5 3.5L13 4"
-                stroke="currentColor"
-                strokeWidth="1.4"
-                strokeLinecap="round"
-                strokeLinejoin="round"
-              />
-            </svg>
-          ) : (
-            <svg viewBox="0 0 16 16" fill="none" className="size-3">
-              <rect
-                x="5"
-                y="5"
-                width="8"
-                height="8"
-                rx="1.5"
-                stroke="currentColor"
-                strokeWidth="1.2"
-              />
-              <path
-                d="M3 11V3.5A.5.5 0 013.5 3H11"
-                stroke="currentColor"
-                strokeWidth="1.2"
-                strokeLinecap="round"
-              />
-            </svg>
-          )}
-        </span>
-      </Button>
-    </div>
-  );
-}
-
-// ── DesktopUpdateCard ─────────────────────────────────────────────────────
-
-// Desktop builds update via electron-updater (a new bundle, swapped in place),
-// not `npm i -g`, so the desktop shows a native action wired to the main
-// process instead of UpdateCard's copyable command. Driven by the auto-update
-// status main pushes over IPC (see hooks/desktop-update).
-function DesktopUpdateCard(props: { update: DesktopUpdate }) {
-  const { status, install } = props.update;
-  const version = "version" in status ? status.version : null;
-
-  const action = (() => {
-    if (status.state === "downloaded") {
-      return (
-        <Button
-          type="button"
-          variant="outline"
-          onClick={install}
-          className="mt-2.5 w-full rounded-lg border-border/60 bg-background/50 px-2.5 py-1.5 text-xs hover:bg-background/80"
-        >
-          Restart to update
-        </Button>
-      );
-    }
-    if (status.state === "downloading") {
-      return (
-        <p className="mt-2.5 text-xs text-muted-foreground tabular-nums">
-          Downloading… {status.percent}%
-        </p>
-      );
-    }
-    if (status.state === "available") {
-      return <p className="mt-2.5 text-xs text-muted-foreground">Preparing update…</p>;
-    }
-    if (status.state === "installing") {
-      return <p className="mt-2.5 text-xs text-muted-foreground">Restarting…</p>;
-    }
-    return null;
-  })();
-
-  return (
-    <div className="mx-2 mb-2 rounded-xl border border-primary/25 bg-primary/[0.06] p-3">
-      <div className="flex items-center gap-2">
-        <div className="flex size-5 shrink-0 items-center justify-center rounded-full bg-primary/15">
-          <svg viewBox="0 0 16 16" fill="none" className="size-3 text-primary">
-            <path
-              d="M8 3v7M5 7l3 3 3-3"
-              stroke="currentColor"
-              strokeWidth="1.4"
-              strokeLinecap="round"
-              strokeLinejoin="round"
-            />
-            <path d="M3 12h10" stroke="currentColor" strokeWidth="1.4" strokeLinecap="round" />
-          </svg>
-        </div>
-        <div className="min-w-0">
-          <p className="text-xs font-semibold text-foreground">Update available</p>
-          {version && <p className="text-sm text-muted-foreground">v{version}</p>}
-        </div>
-      </div>
-      {action}
-    </div>
-  );
-}
 
 // ── NavItem ──────────────────────────────────────────────────────────────
 
@@ -396,15 +151,7 @@ function IntegrationList(props: { pathname: string; onNavigate?: () => void }) {
 
 // ── SidebarContent ───────────────────────────────────────────────────────
 
-function SidebarContent(props: {
-  pathname: string;
-  onNavigate?: () => void;
-  showBrand?: boolean;
-  updateAvailable: boolean;
-  latestVersion: string | null;
-  channel: UpdateChannel;
-  desktopUpdate: DesktopUpdate | null;
-}) {
+function SidebarContent(props: { pathname: string; onNavigate?: () => void; showBrand?: boolean }) {
   const isHome = props.pathname === "/";
   const isSecrets = props.pathname === "/secrets";
   const isPolicies = props.pathname === "/policies";
@@ -467,17 +214,7 @@ function SidebarContent(props: {
         <IntegrationList pathname={props.pathname} onNavigate={props.onNavigate} />
       </nav>
 
-      {/* In the desktop app electron-updater owns updates, so show a native
-          "Restart to update" card and never the npm command. Web and
-          CLI-served installs keep the npm UpdateCard. */}
-      {props.desktopUpdate
-        ? props.desktopUpdate.status.state !== "idle" && (
-            <DesktopUpdateCard update={props.desktopUpdate} />
-          )
-        : props.updateAvailable &&
-          props.latestVersion && (
-            <UpdateCard latestVersion={props.latestVersion} channel={props.channel} />
-          )}
+      <SidebarUpdateCard />
 
       {/* Footer */}
       <div className="shrink-0 border-t border-sidebar-border px-4 py-2.5">
@@ -522,8 +259,6 @@ export function Shell() {
   const pathname = location.pathname;
   const refreshSources = useAtomRefresh(integrationsAtom);
   const refreshTools = useAtomRefresh(toolsAllAtom);
-  const { latestVersion, updateAvailable, channel } = useLatestVersion(VITE_APP_VERSION);
-  const desktopUpdate = useDesktopUpdate();
   const lastPathname = useRef(pathname);
   const [mobileSidebarOpen, setMobileSidebarOpen] = useState(false);
   if (lastPathname.current !== pathname) {
@@ -563,13 +298,7 @@ export function Shell() {
       <CommandPalette />
       {/* Desktop sidebar */}
       <aside className="desktop-macos-sidebar hidden w-52 shrink-0 border-r border-sidebar-border bg-sidebar md:flex md:flex-col lg:w-56">
-        <SidebarContent
-          pathname={pathname}
-          updateAvailable={updateAvailable}
-          latestVersion={latestVersion}
-          channel={channel}
-          desktopUpdate={desktopUpdate}
-        />
+        <SidebarContent pathname={pathname} />
       </aside>
 
       {/* Mobile sidebar overlay */}
@@ -613,10 +342,6 @@ export function Shell() {
               pathname={pathname}
               onNavigate={() => setMobileSidebarOpen(false)}
               showBrand={false}
-              updateAvailable={updateAvailable}
-              latestVersion={latestVersion}
-              channel={channel}
-              desktopUpdate={desktopUpdate}
             />
           </div>
         </div>
